refactor: create openai_test_utils.go with shared test utilities
Created test utilities file and moved all shared mock implementations, helper functions, and test utilities from the large test file. This includes mock API implementation, error handling, token counting, and model info functions.

Part of task R001 from TODO.md, which is the first step in breaking up the large openai_client_test.go file.
